% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.time.subset.vectors.R
\name{check.time.subset.vectors}
\alias{check.time.subset.vectors}
\title{Internal function to check vectors that subset the indicator selection using the time dimension}
\usage{
check.time.subset.vectors(time.vector, vector.name, time, id)
}
\arguments{
\item{time.vector}{A vector containing the user input in e.g. \code{tis_time} or \code{fesis_time}}

\item{vector.name}{The name of argument that the user inputted this vector in. This is just to make error messages more elaborate.}

\item{time}{The time dimension of isatpanel.}

\item{id}{The id dimension of isatpanel.}
}
\value{
Does not return any value but will throw error if something is not correct.
}
\description{
Internal function to check vectors that subset the indicator selection using the time dimension
}
