% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distorttest.boot.R
\name{distorttest.boot}
\alias{distorttest.boot}
\title{Bootstrap Distortion Test Function}
\usage{
distorttest.boot(
  x,
  nboot = 199,
  clean.sample = TRUE,
  parametric = FALSE,
  scale.t.pval = 1,
  parallel = FALSE,
  ncore = detectCores()[1] - 1,
  ...
)
}
\arguments{
\item{x}{isat object (I think)}

\item{nboot}{Number of bootstraps}

\item{clean.sample}{logical}

\item{parametric}{logical. Currently: if parametric then clean.sample = TRUE}

\item{scale.t.pval}{numeric. Default = 1.}

\item{parallel}{logical. Default = FALSE}

\item{ncore}{to be used with parallel}

\item{...}{Further arguments to be passed to ??}
}
\value{

}
\description{
Bootstrap Distortion Test Function
}
\examples{
#' # A quick example with 10 replications
data(Nile)
Nile[10] <- Nile[10]+1000 # to ensure there is an indicator retained
nile <- isat(Nile, sis=FALSE, iis=TRUE, plot=TRUE, t.pval=0.005)

distorttest.boot(nile, nboot = 10)

# Another example with co-variates
dat <- hpdata[,c("GD", "GNPQ", "FSDJ")]
Y <- ts(dat$GD,start = 1959, frequency = 4)
mxreg <- ts(dat[,c("GNPQ","FSDJ")],start = 1959, frequency = 4)
m1 <- isat(y = Y, mc = TRUE, sis = FALSE, iis = TRUE)
m2 <- isat(y = Y, mc = TRUE, sis = FALSE, iis = TRUE, ar = 1)
m3 <- isat(y = Y, mxreg = mxreg, mc = TRUE, sis = FALSE, iis = TRUE)
m4 <- isat(y = Y, mxreg = mxreg, mc = TRUE, sis = FALSE, iis = TRUE, ar = 1)

distorttest(m1)
distorttest(m2)
distorttest(m3)
distorttest(m4)

# bootstrap
db1 <- distorttest.boot(m1, nboot = 10)
db2 <- distorttest.boot(m2, nboot = 10)
db3 <- distorttest.boot(m3, nboot = 10)
db4 <- distorttest.boot(m4, nboot = 10)

db4.clean <- distorttest.boot(m4, nboot = 10, clean.sample = TRUE)
}
