% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distorttest.R
\name{distorttest}
\alias{distorttest}
\title{Distortion Test Function}
\usage{
distorttest(x, coef = "all")
}
\arguments{
\item{x}{An isat object}

\item{coef}{character string. Default is "all"}
}
\value{

}
\description{
Distortion Test Function
}
