% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EU_emissions_road}
\alias{EU_emissions_road}
\title{CO2 Data for EU Road Emissions}
\format{
A data frame with 1550 rows and 13 variables:
\describe{
\item{X}{Index}
\item{country}{Country}
\item{year}{Year}
\item{gdp}{Gross Domestic Product}
\item{pop}{Population}
\item{transport.emissions}{Transport CO2 Emissions}
\item{lgdp}{Log GDP}
\item{lpop}{Log Population}
\item{ltransport.emissions}{Log Transport CO2 Emissions}
\item{const}{Constant}
\item{L1.ltransport.emissions}{Lag 1 Log Transport CO2 Emissions}
\item{L1.lgdp}{Lag 1 Log GDP}
\item{L1.lpop}{Lag 1 Log Population}
}
}
\source{
EDGAR. Data taken from Koch, Naumann, Pretis, Ritter und Schwarz (2022). Nature Energy. \url{https://doi.org/10.1038/s41560-022-01095-6}
}
\usage{
EU_emissions_road
}
\description{
CO2 Data for EU Road Emissions
}
\keyword{datasets}
