% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gets.Fun.R
\name{gets.Fun}
\alias{gets.Fun}
\title{prettify' out$specific.spec:
prettify' out$terminals.results and out$paths:}
\usage{
gets.Fun(
  y,
  x,
  untransformed.residuals = NULL,
  user.estimator = list(name = "ols"),
  gum.result = NULL,
  t.pval = 0.05,
  wald.pval = t.pval,
  do.pet = TRUE,
  ar.LjungB = NULL,
  arch.LjungB = NULL,
  normality.JarqueB = NULL,
  user.diagnostics = NULL,
  gof.function = list(name = "infocrit"),
  gof.method = c("min", "max"),
  keep = NULL,
  include.gum = FALSE,
  include.1cut = FALSE,
  include.empty = FALSE,
  max.paths = NULL,
  turbo = FALSE,
  tol = 1e-07,
  LAPACK = FALSE,
  max.regs = NULL,
  print.searchinfo = TRUE,
  alarm = FALSE
)
}
\description{
prettify' out$specific.spec:
prettify' out$terminals.results and out$paths:
}
