% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_indicator_timings.R
\name{identify_indicator_timings}
\alias{identify_indicator_timings}
\title{Internal function to identify the timing of selected indicators}
\usage{
identify_indicator_timings(object)
}
\arguments{
\item{object}{data.frame}
}
\value{
A list of data.frames
}
\description{
Internal function to identify the timing of selected indicators
}
