% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isatpanel.R
\name{isatpanel}
\alias{isatpanel}
\title{Panel isat function}
\usage{
isatpanel(
  y,
  id,
  time,
  mxreg,
  mxbreak,
  break.method = c("individual"),
  effect = c("individual"),
  iis = FALSE,
  na.remove = TRUE,
  engine = NULL,
  user.estimator = NULL,
  cluster = "individual",
  plm_model = NULL,
  ...
)
}
\arguments{
\item{y}{Year}

\item{id}{Individual}

\item{time}{Time}

\item{mxreg}{The co-variates matrix}

\item{mxbreak}{The break matrix}

\item{break.method}{Break Method}

\item{effect}{Fixed Effect specification}

\item{iis}{use Impulse Indicator Saturation}

\item{na.remove}{remove NAs}

\item{engine}{Function to use}

\item{user.estimator}{Use a user.estimator}

\item{cluster}{cluster Standard Errors at this level}

\item{plm_model}{Type of PLM model (only if engine = "PLM")}

\item{...}{Further arguments to gets::isat}
}
\value{

}
\description{
Panel isat function
}
\examples{
isatpanel(y)
}
