% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isvarcor.isat.R
\name{isvarcor.isat}
\alias{isvarcor.isat}
\title{IIS Consistency Correction for isatobjects (S3 Method)
Consistency correction for estimate of residual variance when using impulse indicator saturation.}
\usage{
isvarcor.isat(isatobject)
}
\arguments{
\item{isatobject}{An isatobject as returned by gets::isat()
The Johansen and Nielsen (2016) impulse-indicator consistency correction for the estimated residual standard deviation.}
}
\value{
An \code{isatobject}  containing the corrected $vcov.mean as well as the original vcov.mean under $vcov.mean.original and additionally the intermediate steps $vcov.mean.cor.1, which is the consistency correction of sigma estimate (affects all regressors) and $vcov.mean.cor.2, which is equal to the new $vcov.mean and represents the additional correction on top of $vcov.mean.cor.1 for the variance of retained regressors (affects only fixed regressors, not impulses).
}
\description{
IIS Consistency Correction for isatobjects (S3 Method)
Consistency correction for estimate of residual variance when using impulse indicator saturation.
}
\examples{
set.seed(123)

########## Setup

y <- rnorm(100, 0, 1)
x <- rnorm(100, 0, 1)
pval = 0.01

is1 <- isat(y, mxreg=x, iis=TRUE, sis=FALSE, t.pval=pval, plot=TRUE)
is1.cor <- isvarcor(is1)

# Modified vcov.mean
is1.cor$vcov.mean
# Equal to is1.cor$vcov.mean.cor.2

# Intermediate Step
is1.cor$vcov.mean.cor.1

# Original vcov.mean
is1.cor$vcov.mean.original
}
