% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.plm.R
\name{logLik.plm}
\alias{logLik.plm}
\title{Log-Likelihood Function for a plm object}
\usage{
\method{logLik}{plm}(object)
}
\arguments{
\item{object}{A plm object}
}
\value{
The Log-Likelihood
}
\description{
Log-Likelihood Function for a plm object
}
