% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pandata_simulated}
\alias{pandata_simulated}
\title{Simulated Panel Data}
\format{
A data frame with 400 rows and 9 variables:
\describe{
\item{country}{A random country}
\item{gdp}{A simulated Gross Domestic Product}
\item{temp}{A simulated variable standing for temperature}
\item{const}{The constant}
\item{country_1}{A dummy for country 1}
\item{country_2}{A dummy for country 2}
\item{country_3}{A dummy for country 3}
\item{country_4}{A dummy for country 4}

...
}
}
\source{
\url{http://github.com/moritzpschwarz/getspanel/}
}
\usage{
pandata_simulated
}
\description{
Simulated Panel Data
}
\keyword{datasets}
