% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plmFun.R
\name{plmFun}
\alias{plmFun}
\title{plm Function to estimate isatpanel}
\usage{
plmFun(y, x, time, id, cluster, effect, model = "pooling", ...)
}
\arguments{
\item{y}{Dependent Variable}

\item{x}{matrix or data.frame of regressors}

\item{time}{Vector of time variable}

\item{id}{Vector of group variable}

\item{cluster}{cluster specification}

\item{effect}{effect specification}

\item{model}{model specification}

\item{...}{Further arguments passed to plm}
}
\value{
A list to be used by gets::isat
}
\description{
plm Function to estimate isatpanel
}
