% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.isatpanel.R
\name{plot.isatpanel}
\alias{plot.isatpanel}
\title{Plotting an isatpanel object}
\usage{
\method{plot}{isatpanel}(
  isatpanelobject,
  max.id.facet = 16,
  facet.scales = "free",
  title = "Panel Saturation",
  ...
)
}
\arguments{
\item{facet.scales}{To be passed to ggplot2::facet_wrap. Default is "free" (i.e. a separate y axis for each panel group/id). Alternatives are: "fixed", "fixed_y", and "fixed_x".}

\item{...}{Further arguments to be passed to ggplot2.}
}
\value{

}
\description{
Plotting an isatpanel object
}
