% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_comp.R
\name{plot_comp}
\alias{plot_comp}
\title{Plotting multiple isatpanel objects for comparison across units}
\usage{
plot_comp(
  mod,
  sign,
  panel = "unit",
  main_text = NULL,
  blanks = TRUE,
  t_range,
  id_list = NULL
)
}
\arguments{
\item{mod}{Dataframe of model objects. At minimum, requires one column with
one isatpanel object wrapped in a list per row and a second column with a unique model description.}

\item{sign}{Type of break to display in plot ("all", negative ("neg"), or positive ("pos")).}

\item{panel}{Whether the results should be displayed by panel cross-section or by model ("model" or "unit")}

\item{main_text}{Plot title (optional)}

\item{blanks}{Boolean - whether to include models in which no break is detected for a panel cross-section.}

\item{t_range}{provide time range of panel from tt:TT.}

\item{id_list}{provide a character vector of unit names to isolate only certain units}
}
\value{
ggplot object that displays multiple isatpanel results by panel-cross section for quick comparison across multiple specifications.
}
\description{
The following function takes a dataframe of isatpanel objects with associated
model names and builds a plot that allows for comparison of detected breaks
by panel cross-section.
Each line should represent a unique model specification-sample pair (no check implemented yet, user must verify).
The function can only handle one outcome/dependent variable at a time.
}
\examples{
library(tidyverse)
library(gridExtra)
library(gets)
library(getspanel)
res <- readRDS(here('data/standard_results_example.RDS')) 
res \%>\% select(is, model) \%>\% plot_comp(t_range = 2000:2021)
# Example isolating specific units
res \%>\% select(is, model) \%>\% plot_comp(t_range = 2000:2021, id_list = c("Argentina", "Bulgaria", "China"))



}
