% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_grid.R
\name{plot_grid}
\alias{plot_grid}
\title{Plotting an isatpanel object}
\usage{
plot_grid(x, title = NULL, ...)
}
\arguments{
\item{x}{An object produced by the isatpanel function}

\item{title}{Plot title. Must be a character vector.}

\item{...}{Further arguments to be passed to ggplot2.}
}
\value{
A ggplot2 plot that plots an 'isatpanel' object and shows all indicators as a grid to give a good and quick overview.
}
\description{
Plotting an isatpanel object
}
