% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_indicators.R
\name{plot_indicators}
\alias{plot_indicators}
\title{Plot isatpanel indicator coefficients}
\usage{
plot_indicators(
  x,
  title = NULL,
  regex_exclude_indicators = NULL,
  zero_line = FALSE,
  include_fixed_effects = FALSE
)
}
\arguments{
\item{x}{An object produced by the isatpanel function}

\item{title}{Plot title. Must be a character vector.}

\item{regex_exclude_indicators}{A regular expression to filter out indicators from the plot. Filter out multiple indicators using "|" (e.g. regex_exclude_indicators = "cfesis|csis". Filter out Unit or Time fixed effects indicators using "id" or "time". Default is NULL, meaning no indicators are excluded.}

\item{zero_line}{Logical. If TRUE, plot a horizontal line at y = 0. Default is FALSE.}

\item{include_fixed_effects}{Logical. If TRUE, fixed effects are included in the plot. Default is FALSE.}
}
\value{
A ggplot2 plot that plots an 'isatpanel' object and shows observed data, the fitted values, and all identified breaks and impulses.
}
\description{
Plot isatpanel indicator coefficients
}
