% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_residuals.isatpanel.R
\name{plot_residuals.isatpanel}
\alias{plot_residuals.isatpanel}
\title{Plot Residuals from 'isatpanel' against OLS}
\usage{
plot_residuals.isatpanel(isatpanelobject)
}
\arguments{
\item{isatpanelobject}{An output from the 'isatpanel' function}
}
\value{

}
\description{
Plot Residuals from 'isatpanel' against OLS
}
\examples{

\dontrun{
data <- pandata_simulated
outcome <- isatpanel(data = data, gdp ~ temp, index = c("country","year"),
effect="twoways",iis=FALSE,fesis=TRUE,t.pval=0.01,engine = "fixest",cluster = "individual")
plot_residuals.isatpanel(outcome)
}

}
