% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_comp.R
\name{plot_unit}
\alias{plot_unit}
\title{This function plots multiple isatpanel objects by single panel cross-section allowing for
comparison across both multiple dependent variables and multiple model specifications.}
\usage{
plot_unit(mod, unit, blanks = TRUE, t_range)
}
\arguments{
\item{mod}{Dataframe of model objects. At minimum, requires one column with
one isatpanel object wrapped in a list per row, a second column with a unique model description,
and a third column with dependent variable names.}

\item{unit}{Exact string identifier of panel cross-section/unit of interest in isatpanel result.}

\item{blanks}{Boolean - whether to include models in which no break is detected for a panel cross-section.}

\item{t_range}{provide time range of panel from tt:TT.}
}
\value{
ggplot object that displays multiple isatpanel results for a single unit for quick comparison across multiple specifications.
}
\description{
This function plots multiple isatpanel objects by single panel cross-section allowing for
comparison across both multiple dependent variables and multiple model specifications.
}
\examples{
res <- readRDS(here('standard_results_example.RDS')) 
res \%>\% select(is, model, dep) \%>\% plot_unit(t_range = 2000:2021, unit = "Peru")


}
