% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.isatpanel.R
\name{print.isatpanel}
\alias{print.isatpanel}
\title{Printing isatpanel results}
\usage{
\method{print}{isatpanel}(x, ...)
}
\arguments{
\item{x}{An isatpanel object.}

\item{...}{Further arguments passed to print}
}
\value{
Print output of the 'isatpanel.result' list element of the 'isatpanel' object.
}
\description{
Printing isatpanel results
}
