% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold.coef_lasso.R
\name{threshold.coef_lasso}
\alias{threshold.coef_lasso}
\title{Run a threshold LASSO based on targeting minimum effect (coefficient) size}
\usage{
threshold.coef_lasso(
  lasso_obj,
  min_coef,
  scale_min_coef = 1,
  absolute = FALSE,
  direction = "negative"
)
}
\arguments{
\item{lasso_obj}{An isatpanel object that was run with engine = 'lasso'}

\item{min_coef}{Numeric.}

\item{scale_min_coef}{Numeric. Scales the minimum coefficient, therefore recommended between 0 and 1.}

\item{direction}{Must be either 'negative' or 'positive'. This determines whether the min_coef is the upper limit (then 'negative' e.g. when min_coef is negative and we want to find coefficients that are even more negative) or the lower limit (then 'positive'). When absolute = TRUE will be ignored.}
}
\value{
A list
}
\description{
Run a threshold LASSO based on targeting minimum effect (coefficient) size
}
