% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_lasso.R
\name{threshold_lasso}
\alias{threshold_lasso}
\title{Run a threshold LASSO based on targeting a number of breaks}
\usage{
threshold_lasso(
  lasso_obj,
  breaks_in_ols_or_lasso = "ols",
  target_neg_breaks,
  target_overall_breaks = NULL,
  plot = TRUE
)
}
\arguments{
\item{lasso_obj}{An isatpanel object that was run with engine = 'lasso'}

\item{breaks_in_ols_or_lasso}{Whether to target the number of breaks in the glmnet (LASSO) object or in the OLS object. The ols object often contains fewer breaks due to collinearity.}

\item{target_neg_breaks}{Integer. Number of negative breaks to target. You can only choose either target_neg_breaks or target_overall_breaks but not both.}

\item{target_overall_breaks}{Integer. Number of breaks to target. You can only choose either target_neg_breaks or target_overall_breaks but not both.}

\item{plot}{Logical. Default is TRUE.}
}
\value{
A list
}
\description{
Run a threshold LASSO based on targeting a number of breaks
}
