% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/within.plm.R
\name{within.plm}
\alias{within.plm}
\title{Use the within transformation from the plm package}
\usage{
\method{within}{plm}(df, effect = "twoways")
}
\arguments{
\item{df}{A data.frame object}

\item{effect}{The fixed effect specification. Values possible: "twoways" (default), "individual", "time", "nested"}
}
\value{
A data.frame object with the transformation complete
}
\description{
Use the within transformation from the plm package
}
