% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/break_uncertainty.R
\name{break_uncertainty}
\alias{break_uncertainty}
\title{Estimate Breakdate Uncertainty}
\usage{
break_uncertainty(x, m = 15, interval = 0.99)
}
\arguments{
\item{x}{An object produced by the isatpanel function}

\item{m}{Maximum range of interval (default is 15 time periods).}

\item{interval}{Approximate level of interval. CI level will be at least > interval. Default 0.99 is a 99\% CI, so the time interval will always be the integer that results in at least > 99\% coverage.}
}
\value{

}
\description{
Estimate Breakdate Uncertainty
}
