% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{compare_models_example}
\alias{compare_models_example}
\title{Example Results for Comparison Plotting}
\format{
A data frame with 3 rows and 5 variables:
\describe{
\item{formula}{Model formula}
\item{country_sample}{Countries included in the sample}
\item{p_val}{P-value of the model}
\item{model}{Unique Model descriptions}
\item{is}{isatpanel objects}
}
}
\source{
\url{https://doi.org/10.1038/s41560-022-01095-6}
}
\usage{
compare_models_example
}
\description{
Example Results for Comparison Plotting
}
\keyword{datasets}
