% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distorttest.boot.R
\name{distorttest.boot}
\alias{distorttest.boot}
\title{Bootstrap Distortion Test Function}
\usage{
distorttest.boot(
  x,
  nboot = 199,
  clean.sample = TRUE,
  parametric = FALSE,
  scale.t.pval = 1,
  parallel = FALSE,
  ncore = detectCores()[1] - 1,
  ...
)
}
\arguments{
\item{x}{isat object (I think)}

\item{nboot}{Number of bootstraps}

\item{clean.sample}{logical}

\item{parametric}{logical. Currently: if parametric then clean.sample = TRUE}

\item{scale.t.pval}{numeric. Default = 1.}

\item{parallel}{logical. Default = FALSE}

\item{ncore}{to be used with parallel}

\item{...}{Further arguments to be passed to ??}
}
\value{

}
\description{
Bootstrap Distortion Test Function
}
