% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EUCO2residential}
\alias{EUCO2residential}
\title{CO2 Data for the EU Residential Sector}
\format{
A data frame with 899 rows and 8 variables:
\describe{
\item{country}{Country}
\item{year}{Year}
\item{lgdp}{Log Gross Domestic Product}
\item{lhdd}{Log Heating Degree Days}
\item{lcdd}{Log Cooling Degree Days}
\item{urban}{Urban Share}
\item{av.rate}{EU Interest Rate}
\item{pop}{Population}
\item{agg.directem}{Aggregated Direct Emissions}
}
}
\source{
IEA
}
\usage{
EUCO2residential
}
\description{
CO2 Data for the EU Residential Sector
}
\keyword{datasets}
