% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_indicators.R
\name{get_indicators}
\alias{get_indicators}
\title{Extract the retained indicators from an \code{isatpanel} object}
\usage{
get_indicators(object, uis_breaks = NULL)
}
\arguments{
\item{object}{An object produced by the isatpanel function.}

\item{uis_breaks}{A string with the names of user-specified indicators.}
}
\description{
Extract the retained indicators from an \code{isatpanel} object
}
