% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isatpanel.R
\name{isatpanel}
\alias{isatpanel}
\title{Panel isat function}
\usage{
isatpanel(
  data = NULL,
  formula = NULL,
  index = NULL,
  effect = c("individual"),
  na.remove = TRUE,
  engine = NULL,
  user.estimator = NULL,
  cluster = "none",
  ar = 0,
  iis = FALSE,
  jiis = FALSE,
  jsis = FALSE,
  fesis = FALSE,
  csis = FALSE,
  cfesis = FALSE,
  csis_var = colnames(mxreg),
  fesis_id = NULL,
  cfesis_var = colnames(mxreg),
  cfesis_id = NULL,
  plot = FALSE,
  plm_model = "within",
  y = NULL,
  id = NULL,
  time = NULL,
  mxreg = NULL,
  ...
)
}
\arguments{
\item{data}{The input data.frame object.}

\item{formula}{Please specify a formula argument. The dependent variable will be the left-most element, separated by a ~ symbol from the remaining regressors. Note the intercept will always be removed, if effect is not "none" - this means that if any fixed effects are specified, the intercept will always be removed.}

\item{index}{Specify the name of the group and time column in the format c("id", "time").}

\item{effect}{Fixed Effect specification. Possible arguments: "twoways", "individual", "time", or "none".}

\item{na.remove}{remove NAs}

\item{engine}{Function to use}

\item{user.estimator}{Use a user.estimator}

\item{cluster}{cluster Standard Errors at this level. Default is "none". Possible values are: "indvidiual", "time", or "twoways".}

\item{ar}{Autoregressive Term to be included. default is 0.}

\item{iis}{use Impulse Indicator Saturation}

\item{jiis}{use Joint Impulse Indicator Saturation (Outliers are common across all units). This is essentially just a time fixed effect, but this allows selection of FE.}

\item{jsis}{use Join Step Indicator Saturation (steps are common across all units)}

\item{fesis}{Fixed Effect Step Indicator Saturation. Constructed by multiplying a constant (1) with group Fixed Effects. Default is FALSE.}

\item{csis}{Coefficient Step Indicator Saturation. Constructed by Default is FALSE.}

\item{cfesis}{Coefficient-Fixed Effect Indicator Saturation. Default is FALSE.}

\item{csis_var}{The csis method can be conducted for all variables or just a subset of them. If you want to use a subset, please specify the column names of the variable in a character vector.}

\item{cfesis_var}{The cfesis method can be conducted for all variables or just a subset of them. If you want to use a subset, please specify the column names of the variable in a character vector.}

\item{plot}{Logical. Should the final object be plotted? Default is FALSE.}

\item{plm_model}{Type of PLM model (only if engine = "PLM")}

\item{y}{Deprecated. The dependent variable. Can be used when data, index, and formula are not specified.}

\item{id}{Deprecated. Can be used when data, index, and formula are not specified. Must be a vector of the grouping variable as a character or factor}

\item{time}{Deprecated. Can be used when data, index, and formula are not specified. Must be a vector of the time variable as an integer or numeric.}

\item{mxreg}{The co-variates matrix}

\item{...}{Further arguments to gets::isat}
}
\value{

}
\description{
Panel isat function
}
\examples{
data(pandata_simulated)
isatpanel(data = pandata_simluated, gdp ~ temp, index = c("country","year"),
effect="twoways",iis=FALSE,fesis=TRUE,t.pval=0.01,engine = "fixest",cluster = "individual")
}
