% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_counterfactual.R
\name{plot_counterfactual}
\alias{plot_counterfactual}
\title{Plot the Counterfactual Path}
\usage{
plot_counterfactual(
  x,
  plus_t = 5,
  facet.scales = "free",
  title = NULL,
  zero_line = TRUE
)
}
\arguments{
\item{x}{An object produced by the isatpanel function}

\item{plus_t}{Number of time periods for the counterfactual to be displayed (default = 5).}

\item{facet.scales}{To be passed to ggplot2::facet_wrap. Default is "free" (i.e. a separate y axis for each panel group/id). Alternatives are: "fixed", "fixed_y", and "fixed_x".}

\item{title}{Plot title. Must be a character vector.}

\item{zero_line}{Plot a horizontal line at y = 0. Default is TRUE.}
}
\value{
A ggplot2 plot that plots an 'isatpanel' object and shows the counterfactuals for each break.
}
\description{
Plot the Counterfactual Path
}
