% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_grid.isatpanel.R
\name{plot_grid}
\alias{plot_grid}
\title{Plotting an isatpanel object}
\usage{
plot_grid(x, max.id.facet = 16, facet.scales = "free", title = NULL, ...)
}
\arguments{
\item{x}{An object produced by the isatpanel function}

\item{title}{Plot title. Must be a character vector.}

\item{...}{Further arguments to be passed to ggplot2.}
}
\description{
Plotting an isatpanel object
}
