% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_residuals.R
\name{plot_residuals}
\alias{plot_residuals}
\title{Plot Residuals from 'isatpanel' against OLS}
\usage{
plot_residuals(isatpanelobject)
}
\arguments{
\item{isatpanelobject}{An output from the 'isatpanel' function}
}
\value{
A ggplot2 plot that plots an 'isatpanel' object and shows the residuals over time in comparison to an OLS model.
}
\description{
Plot Residuals from 'isatpanel' against OLS
}
