% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.isatpanel.R
\name{print.isatpanel}
\alias{print.isatpanel}
\title{Printing isatpanel results}
\usage{
\method{print}{isatpanel}(obj)
}
\arguments{
\item{obj}{An isatpanel object.
#' @export}
}
\description{
Printing isatpanel results
}
